+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityEditComponentPanelTest.java
@@ -29,7 +29,6 @@ import org.junit.jupiter.api.Test;
 
 import javax.swing.JTextField;
 
-import static is.codion.swing.framework.ui.component.EntityComponents.entityComponents;
 import static org.junit.jupiter.api.Assertions.*;
 
 public final class EntityEditComponentPanelTest {
@@ -45,8 +44,6 @@ public final class EntityEditComponentPanelTest {
 	@Test
 	void test() {
 		SwingEntityEditModel editModel = new SwingEntityEditModel(Employee.TYPE, CONNECTION_PROVIDER);
-		assertThrows(IllegalArgumentException.class, () ->
-						new EntityEditComponentPanel(editModel, entityComponents(editModel.entities().definition(Department.TYPE))));
 		EntityEditComponentPanel componentPanel = new EntityEditComponentPanel(editModel);
 		componentPanel.createTextField(Employee.NAME);
 		assertThrows(IllegalStateException.class, () -> componentPanel.createTextField(Employee.NAME));